Project with Gemini/requirements.txt
fastapi==0.110.0
uvicorn==0.25.0
httpx==0.25.2
pydantic==2.6.1
paho-mqtt==2.2.1
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
google-generativeai==0.3.1 